No longer using public methods returing pointers in the atom style to access per-shape arrays. Per-shape arrays are not in the atom scope, but the memory for them are allocated within the atom style itself. The result is that they can be accessed via atom->shape_array.
Now referring to "shapes" instead of "shtypes" to avoid the confusion with the LAMMPS type.

Added flags to the pair_style so that it can switch between using the points of Gaussian quadrature and the patch approach. Both are flawed (contact is not equal and opposite due to the non-equal distance between surface points).