Removed all angled brackets in include statements
Had to add "quat" to fields_border_vel in atom_vec_spherharm in order for the quaternion of atoms on different processors to be communicated

compute_efunction_spherharm needed to be updated as it was still loading mass based on types rather than rmass. This was resulting in a mass=0 and a divion by zero leading to NANS. This has been fixed.

Tidying up the pair_style:
Deleting the bisection gradient method as this is WIP and will only serve to confuse at present. Knock on effect of removing calc_nearest_point and skew_lines_check.

Removing get_contact_point_bounds as this seems not to be called anywhere. May be called in commented out code but that remains to be seen.